@@ -23,51 +23,51 @@
     "test:e2e": "jest --config ./test/jest-e2e.json"
   },
   "dependencies": {
-    "@nestjs/common": "^8.1.2",
-    "@nestjs/core": "^8.1.2",
-    "@nestjs/platform-express": "^8.1.2",
-    "@nestjs/sequelize": "^8.0.0",
-    "@nestjs/terminus": "^8.0.1",
-    "aws-sdk": "^2.1177.0",
+    "@nestjs/common": "^9.0.7",
+    "@nestjs/core": "^9.0.7",
+    "@nestjs/platform-express": "^9.0.7",
+    "@nestjs/sequelize": "^9.0.0",
+    "@nestjs/terminus": "^9.0.0",
+    "aws-sdk": "^2.1184.0",
     "convict": "^6.2.3",
-    "express-session": "^1.17.2",
-    "helmet": "^4.6.0",
+    "express-session": "^1.17.3",
+    "helmet": "^5.1.1",
     "nodemailer": "^6.7.7",
     "otplib": "^12.0.1",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
-    "rxjs": "^7.4.0",
+    "rxjs": "^7.5.6",
     "sequelize": "^6.21.3",
-    "sequelize-typescript": "^2.1.1",
-    "sqlite3": "^5.0.9"
+    "sequelize-typescript": "^2.1.3",
+    "sqlite3": "^5.0.10"
   },
   "devDependencies": {
-    "@nestjs/cli": "^8.1.4",
-    "@nestjs/schematics": "^8.0.4",
-    "@nestjs/testing": "^8.1.2",
+    "@nestjs/cli": "^9.0.0",
+    "@nestjs/schematics": "^9.0.1",
+    "@nestjs/testing": "^9.0.7",
     "@types/convict": "^6.1.1",
     "@types/express": "^4.17.11",
-    "@types/express-session": "^1.17.4",
+    "@types/express-session": "^1.17.5",
     "@types/helmet": "^4.0.0",
-    "@types/jest": "^27.0.2",
+    "@types/jest": "^28.1.6",
     "@types/node": "^16.11.6",
     "@types/nodemailer": "^6.4.4",
-    "@types/sequelize": "^4.28.10",
-    "@types/supertest": "^2.0.10",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "eslint": "^7.22.0",
-    "eslint-config-prettier": "^8.1.0",
+    "@types/sequelize": "^4.28.14",
+    "@types/supertest": "^2.0.12",
+    "@typescript-eslint/eslint-plugin": "^5.31.0",
+    "@typescript-eslint/parser": "^5.31.0",
+    "eslint": "^8.20.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "jest": "^26.6.3",
-    "lint-staged": "^11.2.6",
-    "prettier": "^2.4.1",
-    "supertest": "^6.1.6",
-    "ts-jest": "^26.5.4",
-    "ts-loader": "^9.2.6",
+    "jest": "^28.1.3",
+    "lint-staged": "^13.0.3",
+    "prettier": "^2.7.1",
+    "supertest": "^6.2.4",
+    "ts-jest": "^28.0.7",
+    "ts-loader": "^9.3.1",
     "ts-node": "^10.9.1",
-    "tsconfig-paths": "^3.11.0",
-    "typescript": "^4.4.4"
+    "tsconfig-paths": "^4.0.0",
+    "typescript": "^4.7.4"
   },
   "lint-staged": {
     "**/*.(js|jsx|ts|tsx)": [
