@@ -1,30 +1,11 @@
-import { Redirect, Route, RouteProps } from 'react-router-dom'
+import { Navigate } from 'react-router-dom'
 
-import { LOGIN_ROUTE } from '~constants/routes'
-
+import { routes } from '~constants/routes'
 import { useAuth } from '~features/auth'
 
-export const PrivateRoute = ({
-  children,
-  ...rest
-}: Omit<RouteProps, 'render'>): JSX.Element => {
+export const PrivateRoute = ({ children }: { children: JSX.Element }) => {
   const { isAuthenticated } = useAuth()
 
-  return (
-    <Route
-      {...rest}
-      render={({ location }) => {
-        return isAuthenticated ? (
-          children
-        ) : (
-          <Redirect
-            to={{
-              pathname: LOGIN_ROUTE,
-              state: { from: location },
-            }}
-          />
-        )
-      }}
-    />
-  )
+  if (!isAuthenticated) return <Navigate to={routes.login} />
+  return children
 }
