@@ -1,2 +1,5 @@
-export const ROOT_ROUTE = '/'
-export const LOGIN_ROUTE = '/login'
+export const routes = {
+  index: '/',
+  login: '/login',
+  health: '/health',
+}
